Improve modeling of meal/dessert prep
r3952
2017-04-26a

Note: this commit is a part of project #5 at https://github.com/mars-sim/mars-sim/projects/5

# NEW
1. Add how the culinary skill as well as the current performance 
   score of the cook would affect the meal/dessert quality index.
   - Insert a new param to addWork() in PreparingDessert and Cooking.
   - Add contribution of a person/robot's performance score to the 
     meal/dessert's quality.
2. Add cleanUpKitchen() in PreparingDessert.


# CHANGES
1. Modify areAllIngredientsAvailable() in Cooking to filter checking 
   for the presence of the ingredient id 0, 1, and 2. 
   Note : ingredients beyond id 2 are optional.
2. Limit the max and min value of cleanliness in Cooking.
3. Remove calling setChef() in CookMeal and prepareDessert's constructor.
4. Rename useWater() to consumeWater() and add randomness in amount of 
   water usage in PreparingDessert.
5. Increase PREPARE_DESSERT_WORK_REQUIRED from 2 to 3 millisols 
   in PreparingDessert.
6. Convert from int to double in representing dessert quality index in
   various class.      
7. Rename cleanUp() to finishUp() in PreparingDessert and Cooking.

+++ mars-sim-core/src/main/java/org/mars_sim/msp/core/structure/building/function/cooking/PreparedDessert.java
@@ -27,12 +27,12 @@ implements Serializable {
 	private static final double SHELF_LIFE = 150D;
 
 	// Data members
-	private int quality;
+	private double quality;
 	private MarsClock expirationTime;
 
 	// 2014-11-28 Added name
 	private String name;
-	
+
 	private String producerName;
 	private String consumerName;
 	private PreparingDessert kitchen;
@@ -43,23 +43,23 @@ implements Serializable {
 	 * @param quality the quality of the dessert
 	 * @param creationTime the time the dessert was cooked.
 	 */
-	public PreparedDessert(String name, int quality, double dryMass, MarsClock creationTime, String producerName, PreparingDessert kitchen) {
+	public PreparedDessert(String name, double quality, double dryMass, MarsClock creationTime, String producerName, PreparingDessert kitchen) {
 		this.quality = quality;
 		this.name = name;
 		this.dryMass = dryMass;
 		expirationTime = (MarsClock) creationTime.clone();
 		expirationTime.addTime(SHELF_LIFE);
 		this.producerName = producerName;
-		this.kitchen = kitchen; 
+		this.kitchen = kitchen;
 	}
-	
+
 	public PreparedDessert(PreparedDessert preparedDessert, String consumerName) {
 		this.quality = preparedDessert.quality;
 		this.name = preparedDessert.name;
 		this.expirationTime = preparedDessert.expirationTime;
 		this.consumerName = consumerName;
 	}
-	
+
 	// 2014-11-28 Added getName()
 	public String getName() {
 		return name;
@@ -68,7 +68,7 @@ implements Serializable {
 	 * Gets the quality of the meal.
 	 * @return quality
 	 */
-	public int getQuality() {
+	public double getQuality() {
 		return quality;
 	}
 
@@ -79,7 +79,7 @@ implements Serializable {
 	public double getDryMass() {
 		return dryMass;
 	}
-	
+
 	/**
 	 * Gets the expiration time of the meal.
 	 * @return expiration time
